Feat: expandableText now works via html classes
Expandable Text can be implemented by adding the class 'expandable-text' to any element containing text.

Via JS the text will automatically be truncated and a button with read more will be added. This implementation also ensures that,
if JS does not load, the content is still available.

With the html attribute 'displaylimit' the point at which the text gets truncated can be set, the default value is 220.