Test: Added CI testing
It will run python unit tests with pytest and e2e tests with cypress.

The build stage builds Nebula for the unit tests. 
For the e2e test a full clean ubuntu image is created in which all
 of the prerequisites are installed.

Reports are created and are available in the CI-CD Pipeline.

Cypress will save a video of the test which is deleted after 1 day.